% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind_eqsym.R
\name{bind_sym}
\alias{bind_sym}
\title{helper for converting ENSG to symbols, cbinding symbol and chromosome name for the genes to the input data.frame-like entity}
\usage{
bind_sym(
  dflike,
  gene_resource = genes_v75,
  ensgvbl = "trait",
  output_symvbl = "symbol",
  output_chromvbl = "gchrom"
)
}
\arguments{
\item{dflike}{a data.frame-like entity with a column of ENSG gene identifiers; column
name must be the value of \code{bind_sym} parameter \code{ensgvbl}}

\item{gene_resource}{a data.frame with columns \code{gene_name} and \code{seqnames},
defaulting to \code{gwaslake::genes_v75}}

\item{ensgvbl}{character(1) the column of \code{dflike} holding the ENSG ids}

\item{output_symvbl}{character(1) name to be used for symbol in output data.frame, defaults to \code{symbol}}

\item{output_chromvbl}{character(1) name to be used for chromosome in output data.frame, defaults to \code{gchrom}}
}
\description{
helper for converting ENSG to symbols, cbinding symbol and chromosome name for the genes to the input data.frame-like entity
}
\examples{
head(little_eq)
head(bind_sym(little_eq))
}
