% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchwork.R
\name{browse_ieu}
\alias{browse_ieu}
\title{a simple app supporting filtering and exploration of MRC IEU GWAS batches and associated studies}
\usage{
browse_ieu(dat = NULL)
}
\arguments{
\item{dat}{a data.frame-like thing expected to be like the result of ieugwasr::gwasinfo(); if NULL,
the output of gwasinfo() will be used}
}
\value{
just invokes a shiny app at this point, no return
}
\description{
a simple app supporting filtering and exploration of MRC IEU GWAS batches and associated studies
}
