% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{genes_v75}
\alias{genes_v75}
\title{a data.frame of genes(EnsDb.Hsapiens.v75) with build appended}
\format{
data.frame
}
\usage{
genes_v75
}
\description{
a data.frame of genes(EnsDb.Hsapiens.v75) with build appended
}
\examples{
head(genes_v75,3)
}
\keyword{datasets}
