% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variants.R
\name{genesym_to_string}
\alias{genesym_to_string}
\title{create a "chr:start-end" string for associations()}
\usage{
genesym_to_string(genesym, radius = 0, addr_src = gwaslake::genes_v75)
}
\arguments{
\item{genesym}{character(1)}

\item{radius}{numeric(1) flanking range, defaults to zero}

\item{addr_src}{data.frame defaults to \code{genes_v75}}
}
\description{
create a "chr:start-end" string for associations()
}
\examples{
genesym_to_string("YY1")
}
