% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gwidf_2021_01_30}
\alias{gwidf_2021_01_30}
\title{a snapshot of output of ieugwasr::gwasinfo()}
\format{
data.frame
}
\usage{
gwidf_2021_01_30
}
\description{
a snapshot of output of ieugwasr::gwasinfo()
}
\note{
Includes 'batchcode' computed by code commented out in example
}
\examples{
# batchcode = strsplit(gwidf_2021_01_30$id, "-")
# batchcode = sapply(batchcode,function(x) paste0(x[1], "-", x[2]))
head(gwidf_2021_01_30)
}
\keyword{datasets}
