% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/broker.R
\name{ieutab_to_cravat}
\alias{ieutab_to_cravat}
\title{format the variant location, content and identity information from MRC IEU API tables for OpenCRAVAT}
\usage{
ieutab_to_cravat(ieutab, chain = NULL, genome = NULL)
}
\arguments{
\item{ieutab}{a data.frame-like entity with \code{chr}, \code{position}, \code{ea} (to be regarded}

\item{chain}{a liftOver chain structure as defined in rtracklayer}

\item{genome}{character(1) label, if present, attached to output as a column \code{build}
as ALT allele and \code{nea} to be regarded as REF allele}
}
\description{
format the variant location, content and identity information from MRC IEU API tables for OpenCRAVAT
}
\note{
If a liftOver chain is provided, some input records may be lost.  We attempt
to keep allele content information with each locus.
}
\examples{
# Here we assume it is desired to give hg38 variant positions to cravat
if (requireNamespace("ieugwasr")) {
  pw = try(ieugwasr::phewas("17:37000000-37020000"))
  if (!inherits(pw, "try-error")) {
    octab = ieutab_to_cravat(pw, chain=gwaslake::ch19to38, genome="hg38" )
    print(head(octab))
    }
  aa = try(ieugwasr::associations("rs6060535", id="ukb-b-10787"))
  if (!inherits(aa, "try-error")) {
    octab2 = ieutab_to_cravat(aa, chain=gwaslake::ch19to38, genome="hg38")
    head(octab2)
    }
}
}
