% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variants.R
\name{ieutab_to_GRanges}
\alias{ieutab_to_GRanges}
\title{IEU tables conversion to GRanges}
\usage{
ieutab_to_GRanges(ieutab, build = "GRCh37")
}
\arguments{
\item{ieutab}{data.frame-like entity as returned by e.g., variant_rsid, associations, phewas}

\item{build}{character(1) defaults to "GRCh37"}
}
\description{
IEU tables conversion to GRanges
}
\note{
assumes SNP; renames 'pos' to 'position' if necessary
}
\examples{
if (requireNamespace("ieugwasr")) {
ex = ieugwasr::variants_rsid("rs6060535")
ieutab_to_GRanges(ex)
}
}
