% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{little_eq}
\alias{little_eq}
\title{a data.frame of eQTL results for testing}
\format{
data.frame
}
\usage{
little_eq
}
\description{
a data.frame of eQTL results for testing
}
\examples{
head(little_eq)
}
\keyword{datasets}
