% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manh.R
\name{manhattanPlot}
\alias{manhattanPlot}
\title{produce a manhattan plot for a given study, based on specific coordinates or gene symbol with radius for flanking region}
\usage{
manhattanPlot(
  studyID = "ubm-a-524",
  symbol = "BCAN",
  start = NULL,
  stop = NULL,
  chromosome = NULL,
  radius = 0,
  datasource = gwidf_2021_01_30
)
}
\arguments{
\item{studyID}{character(1) id of study in ieugwasr::gwasinfo() catalog, defaults to "ubm-a-524", which is
a study described in Elliott et al Nature October 2018}

\item{symbol}{a gene symbol (to be used with \code{genesym_to_string}, implicitly); if absent, chromosome, start and end must be supplied, defaults to 'BCAN'}

\item{start}{chromosomal address to start region (will have radius subtracted if supplied)}

\item{stop}{chromosomal address to end region (will have radius added if supplied)}

\item{chromosome}{name of chromosome in Ensembl nomenclature (e.g., 1, 2, ...)}

\item{radius}{numeric default to zero, specifies flanking region}

\item{datasource}{data.frame that includes study identifier \code{id} and \code{trait} for labeling}
}
\description{
produce a manhattan plot for a given study, based on specific coordinates or gene symbol with radius for flanking region
}
\examples{
manhattanPlot(radius=500000)
}
