% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mapped_traits_demo}
\alias{mapped_traits_demo}
\title{a subset of trait labels mapped to Disease Ontology}
\format{
named character vector
}
\usage{
mapped_traits_demo
}
\description{
a subset of trait labels mapped to Disease Ontology
}
\note{
Uses ontoProc from Bioc 3.12.
}
\examples{
head(mapped_traits_demo)
}
\keyword{datasets}
