% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_gwas.R
\name{survey_app}
\alias{survey_app}
\title{simple app to help survey diverse phenotypes}
\usage{
survey_app()
}
\description{
simple app to help survey diverse phenotypes
}
\note{
Will try to find hits at 1e-8, if none, increases to 1e-6.
}
