% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_gwas.R
\name{survey_gwas}
\alias{survey_gwas}
\title{plot sample size and other information about studies in IEU GWAS ecosystem}
\usage{
survey_gwas(
  phrases,
  datasource = gwaslake::gwidf_2021_01_30,
  title_pref = "IEU GWAS",
  title_text = paste(sQuote(phrases), collapse = ", "),
  xlab = "N controls",
  ylab = "N cases",
  ...
)
}
\arguments{
\item{phrases}{character() vector to grep with ignore.case=TRUE in \code{trait} field, unique matching rows
will be collected}

\item{datasource}{data.frame with \code{trait} as a column, defaults to
gwaslake::gwidf_2021_01_30; result of ieugwasr::gwasinfo() is intended input}

\item{title_pref}{character(1) defaults to "IEU GWAS"}

\item{title_text}{character(1) defaults to a paste of phrases with collapse=", "}

\item{xlab}{character(1) x axis label defaults to "N controls"}

\item{ylab}{character(1) x axis label defaults to "N cases"}

\item{\dots}{passed to grep}
}
\description{
plot sample size and other information about studies in IEU GWAS ecosystem
}
\note{
returns ggplot, intended for ggplotly
}
\examples{
requireNamespace("ggplot2")
sg = survey_gwas(c("asthma", "asthmatic")) + 
 ggplot2::theme(text=ggplot2::element_text(size=16))
if (interactive()) {
 requireNamespace("plotly")
 plotly::ggplotly(sg)
}
}
